pyproject.toml
[project]
name = "steel-city-red-teamning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.11.0",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
]
